Create ViewModel and add navigation bar
This navigation bar appears and disappears on double tap. This is mapped
to a double tap gesture since its easy to show and doesn't interfere
with the overall webpage experience.

This utilizes a UIKit gesture handler which changes a boolean in the
ViewModel.

Also create a ViewModel which handles the buisness logic of the WebView.

Signed-off-by: kingbri <bdashore3@gmail.com>