Settings: Add request desktop/mobile site toggle
This changes the default WebView user agent to either one from
a Mac or an iPhone depending on the device you're running on.

User agent changing switches the website layout from desktop to
mobile (if the website supports it).

This toggle persists on app relaunches and anomalies are removed
by clearing the cache on app load.

Accounts are not removed on cache clear
- Test: Sign into google, close app, reopen app, still signed in.

Signed-off-by: kingbri <bdashore3@gmail.com>