Navigation: Add loading and error handling
Add a ProgressView for a loading animation on page load. This is
triggered with the refresh button and navigation delegates.

Also add a textbox for any errors that show up during page loading.
This would be helpful for pages such as mangadex which explains
that a timeout occured.

Finally, add some helpful comments along with removing the white
background when loading content.

Signed-off-by: kingbri <bdashore3@gmail.com>