Mac: Fix URL being set when opening the settings
This is due to a bug in MacCatalyst when opening a swiftUI form
and scrolling.

This behavior causes the first textbox to become active, or "focused".
Exhaustive testing found little to no solutions for this issue. The best
workaround is to make Mac users hit the "enter" key for setting the
default URL.

Signed-off-by: kingbri <bdashore3@gmail.com>