WebView: Ignore error -1022
Error -1022 is something called an App transport security policy error.
This is to enforce secure connections to websites and doubles as
a helpful popup ad blocker.

However, the error handler in the browser made this error show over
and over when using 9anime. Add an if statement to ignore that error
and display all other ones.

More annoying errors will be ignored on request.

Signed-off-by: kingbri <bdashore3@gmail.com>