Asobi: Add Coredata and bookmarks
Add a Coredata model which saves values for bookmarks and history.
Currently, bookmarks are implemented for iPhone devices.

Bookmarks are saved hyperlinks to a webpage. They can be edited
using a ContextMenu on iOS 14 and swipe actions on iOS 15.

These can also be deleted with the help of EditMode.

The overall navigation bar has been cleaned up to tuck About in
the settings menu and expose a new menu for combined bookmarks and history.

NOTE: This does not work on iPad/macOS yet.

Signed-off-by: kingbri <bdashore3@gmail.com>