Navigation: Show popup errors and fix navigation
The current WebView wasn't extending to the entirety of the screen.
Fix this by ignoring safe area on the webview for the bottom
edge of the screen.

Make errors popup and disappear after a period of time (5 seconds)
so that the user doesn't have to reload the webpage.

Also add an isNotch boolean to UIDevice to check if padding
is needed on the navigation bar or not.

Signed-off-by: kingbri <bdashore3@gmail.com>