Navigation: Use a ViewModel
There were too many bindings and reliances on PresentationMode in this
app since it relies on sheets. The sheets also depended on the
NavigationBarView which were closed out when the navigation bar disappeared
(thanks to the auto-hide setting)

Therefore, create a global ViewModel to control sheet state from
any view. This allows us to use the common ListRowLinkView again
since the dismissal of a sheet doesn't require a binding.

Signed-off-by: kingbri <bdashore3@gmail.com>