Views: Add keyboard shortcuts for MacCatalyst
Use escape to navigate away from the sheet. In EditBookmarkView,
enter saves the bookmark that you're editing.

For some reason, including a color picker and toggle in the same
form section causes keyboard shortcuts to break on macOS.

A bug report has been filed. Users will not be able to exit
the settings menu without clicking "done" on macOS.

Signed-off-by: kingbri <bdashore3@gmail.com>