WebView: Enable full AirPlay support
AirPlay is a bit weird when working with a WebView. Some websites
only show the AirPlay button on a video player if Safari is used.

Therefore, make the default user agent some form of safari to allow
AirPlay options to be shown.

Also, don't automatically use AVPlayer when playing a video so a user
can click on the AirPlay options.

This may be a toggleable option in the future due to possible video
player ads that can show up with inline play.

Signed-off-by: kingbri <bdashore3@gmail.com>