Asobi: Add history
This commit adds browser history to Asobi. All duplicate URL entries
are overwritten and separated by day.

There is also a method to clear browser data for the following intervals:
- Past day
- Past week
- Past 4 weeks (since a month is a variable timeperiod)
- All time

Also add a bunch of cleanup changes for cosmetic purposes.

The library view was now changed to a grouped list style instead
of an insetGrouped one.

Signed-off-by: kingbri <bdashore3@gmail.com>