HistoryView: Fix actions button
There is a bug in iOS 14 that prevents NavigationView toolbars
from showing action sheets.

To fix this, make iOS 14 and less show an inlined actions button
to properly show the action sheet.

Also, edit the swipeActions conditional in BookmarkView to use
the UIDevice rather than the process OS version.

The process OS version on a MacBook will eventually become 15, the
UIDevice method makes sure the swipe actions will always show
on an iPhone/iPad.

The same logic is applied to the inline buttton.

Signed-off-by: kingbri <bdashore3@gmail.com>