WebView: Add download support
This framework adds the ability to download files when a download
URL is opened.

A download will also fire if there's no HTML content on the webpage,
this is just the behavior of WKWebView since there's no true
way to detect a download without specifying the MIME types to look for.

This framework allows for single downloading only. This means that
one download can occur at a time. If a user tries to start another
download, an alert will pop up saying to cancel the current download.

A download manager along with a save to files button may be added
in the future. It isn't too difficult to add, but the framework
is currently optimized for single downloads.

Signed-off-by: kingbri <bdashore3@gmail.com>