Downloads: Migrate to swift concurrency syntax
With AlamoFire's release for swift concurrency, migrate the download
handler to be cleaner and more readable.

The DownloadProgress timer needed a slight revamp to prevent race
conditions inside an async context (since downloadProgress uses a
for loop now).

Add timer properties in an actor to provide synchronous setting
in an asynchronous loop.

Signed-off-by: kingbri <bdashore3@gmail.com>