WebView: Remove complexity in background color changes
Set the background of a webpage on load. Ignore all safe areas except
the top one in portrait mode so the background is displayed everywhere
but the status bar.

The "Set status bar background to page color" option has been removed
with this change since some background colors can conflict with statusbar
text.

Signed-off-by: kingbri <bdashore3@gmail.com>