WebViewModel: Fix zooming bugs with iOS 14
The orientation change bug is worse on iOS 14, Apple must have
fixed parts of WebKit in iOS 15, so the tolerance to force zoom
out on orientation change is very close to 100. However, with iOS 14,
the tolerance is between 100-200% zoomed in.

This is not an acceptable tolerance range and requires a lot of unnecessary
code to implement. Therefore, on iOS 14, on an orientation change,
the webpage gets zoomed out completely no matter how zoomed in
the user is.

This may be changed in the future.

Also, only enable pinch to zoom for every page if the user is on an iPad
or iPhone

Signed-off-by: kingbri <bdashore3@gmail.com>