WebView: Properly fix zooming
This overhauls the orientation based syntax and uses UI Delegates
to primarily handle changes in orientation and zoom offsets.

There are two zoom calls fired on an orientation change and only
one was caught by the previous orienation handler. The scroll delegate
catches both of the zoom calls and adjusts accordingly.

In addition, the zoom scale doesn't persist between orientation changes.
Save the previous zoom scale if a user zoomed in and apply it on
orientation change.

Signed-off-by: kingbri <bdashore3@gmail.com>