Settings: Add manual theme selection
Selecting dark or light themes may be pertinent for users down the road.

However, there is an issue when following the system theme which
requires the view to be dismissed and refreshed with the new theme
color.

This is currently a workaround that'll stay since instant changes
of light/dark theme can be jarring to the eyes.

Signed-off-by: kingbri <bdashore3@gmail.com>