Library: Use a UUID id for lists
If a user's bookmarks or history becomes big, it can be bad for performance
when a user updates bookmarks or history.

Using a UUID for the id will trigger a list replace when the list
is updated. This will allow the list to be redrawn as opposed
to being compared if there are a large amount of entries.

Signed-off-by: kingbri <bdashore3@gmail.com>