Asobi: Fix manual URL loading
Asobi now supports other URL schemes besides http/https.

A great way to check if a URL contains a scheme is to check for the
:// characters in succession. One of these substrings means a URL
is present and the user is trying to navigate to it.

However, the WebView may get confused with the URL scheme, so always
allow malformed URL calls through so they can be propegated to a
user-facing error.

Signed-off-by: kingbri <bdashore3@gmail.com>