WebView: Preliminary support for find in page
This commit allows a user to find content in a page. The search
is manual to save on resources. This requires the user to click
the search button when finding an item on the webpage.

The currently selected item is highlighted and can be changed by
clicking on the chevron icons.

Exiting out of the find in page menu clears the search and webpage
highlights.

Also updated the navigation bar logic since the Find in page searchbar
was not being carried by the iOS virtual keyboard.

Signed-off-by: kingbri <bdashore3@gmail.com>