Downloads: Fix UserDefaults bug
Using @AppStorage property wrappers outside of a view allow
for read-only values. Writing to this variable causes the source
of truth to be scoped to what the viewmodel wrote.

The best workaround is to use UserDefaults.standard whenever
a value is being set and use the @AppStorage value to read the
data (since optionals don't have to be dealt with).

Signed-off-by: kingbri <bdashore3@gmail.com>