MacCatalyst: Remove blur when minimized
Applying blur across views causes the MacCatalyst version of Asobi
to crash. This is due to how EnvironmentObject works. The EnvironmentObject
isn't loaded until the view itself is called (aka the settings sheet
shows up).

This can be mitigated by using the onAppear modifier to set a local
state variable, but this is too much refactoring for a blur feature
which isn't that useful on the MacOS side of things.

Since blur isn't that necessary of a feature, purge it from the MacCatalyst
build of Asobi.

Signed-off-by: kingbri <bdashore3@gmail.com>