Settings: Fix crash with Done button (iOS 14)
iOS 14 has an issue where an environmentObject isn't loaded when
the view is instantiated. It turns out when pressing the "done"
button, Asobi crashes since the Navigation ViewModel can't be found.

This doesn't occur on iOS 15, presumably because SwiftUI 3 has
fixed this issue.

Pass the environment object into the SettingsView if the user is on
macOS or iOS 14.

Resolves #37

Signed-off-by: kingbri <bdashore3@gmail.com>