Asobi: Batch small feature commit
1. There is now a preference to playback video in fullscreen by default.
Playing videos in fullscreen uses Apple's native video controls. This
setting requires an app restart since the preference is set on
WebView configuration.

2. Do not clear the website caches by default. This change was introduced
back when Asobi was just a Cubari reader and required a fresh
launch every time the app opened. If a user wants to preserve all
account logins, the new preference should be off.

3. Add an action to force refresh a page. Sometimes a page can break
and the pull-to-refresh functionality can break as well. On iPhones,
the refresh button doesn't exist in the navbar, so add something
in actions for all devices.

Signed-off-by: kingbri <bdashore3@gmail.com>