Settings: Add statusbar customization
This allows full customization of the statusbar for Asobi.

There are two sets of settings for the statusbar:
1. Behavior: This setting allows the user to change the statusbar pinning.
   There are 3 options:
   - pinned: Statusbar always shown
   - partially hidden: Statubar hidden when navbar isn't showing
   - hidden: Statusbar not shown to the user

2. Color: This setting allows the user to customize the color of the
   statusbar. There are 3 options:
   - Theme: Adapts to light/dark theme
   - Automatic: Adapts to the webpage's background color
   - Custom: The user can select the color and that will persist

To implement this (especially the color), a drop-in replacement for
the HostingViewController is required. This replacement is a StateObject
which observes changes in settings and adjusts them accordingly.

The easiest way to grab the rootView (to set the HostingViewController)
is to use Introspect on the MainView of the app.

But, reimplementing the HostingViewController also broke SwiftUI's
statusbar modifier, so this also had to be implemented in the new
HostingViewController.

I may change this to use modifiers in the future, but using onChange
works fine for now.

Signed-off-by: kingbri <bdashore3@gmail.com>