Asobi: Split code
Decouple the JS loading code from the WebModel because it looked
really ugly.

In addition, separate all inlined JS into their own scripts with a
loading system. This will be made dynamic in the future.

Finally, change the error toast to sleep a second before showing
due to the chaos of WebView loading interfering with toast presentation
on app init.

Signed-off-by: kingbri <bdashore3@gmail.com>