Settings: Persist custom statusbar color and add accent option
This change allows for the user's custom statusbar color to be preserved
across style changes and also adds the option to change the statusbar
color to the app's accent color.

In addition, all AppStorage calls in models have been set to a private
protection level to keep AppStorage as the single source of truth.

Signed-off-by: kingbri <bdashore3@gmail.com>