iCloud: Add preference switch and fixes
A toggle for syncing has been added. When the switch is toggled to on,
any data is pushed to iCloud and handled there.

The toggle currently requires an app restart due to crashes.

This also fixes how history is presented with iCloud syncing since
some history entries may be split and will show weird results.

Signed-off-by: kingbri <bdashore3@gmail.com>