Settings: Add home indicator and inset options
These settings only show up on notched iDevices.

Notch: There is now the option to gray out the home indicator when
the navigation bar is hidden. This will allow users to not have
a large homescreen pill on the screen when insets are off.

Inset: The bottom inset used to be removed by default to provide
continuity to the webview. However, the bottom inset is sometimes
required especially for sites with a bottom navigation bar.

Signed-off-by: kingbri <bdashore3@gmail.com>