Text: Add ability to clear all text in fields
When editing anything in Asobi, it's a good idea to have a button
which easily clears text in the field. This is in UIKit by default,
so introspect is used to port this functionality to SwiftUI.

Also make the disabledAppearance modifier actually disable the view
along with giving the grayed out appearance.

Signed-off-by: kingbri <bdashore3@gmail.com>