History: Add repair function
There is an anomaly where some history entries do not have a parent
history to display inside Asobi's library view.

This will interfere with stateful bookmarks and load last history
entry for the default URL.

The repair history function re-associates these zombie entries
with a parent history. If the history for that date doesn't exist,
create a new one.

Signed-off-by: kingbri <bdashore3@gmail.com>