Asobi: Add global keyboard commands for MacOS
Commands for find in page and zooming are now globally available in
MacOS.

Zoom is activated via cmd + shift + plus or minus. I'm working
on making this cmd + plus or minus but there is a weird limitation
preventing me from making this shortcut work.

Zooming on webpages is also possible using the trackpad pinch
gesture.

WebModel has been moved to the App file for top level processing
of commands.

Signed-off-by: kingbri <bdashore3@gmail.com>