Library: Fix editing bookmarks with navigation
The EditBookmarkView call used a deprecated NavigationLink initializer
as a workaround for the toolbar not showing up when the edit view
disappears.

A more permanent solution is to expose the viewWillDissappear API
and link up these views manually.

onDisappear could have been used, but that resulted in a jarring
transition for the toolbar's state to be refreshed.

Signed-off-by: kingbri <bdashore3@gmail.com>