Navigation: Fix context menu disappearance with auto-hide
Auto-hiding the navigation bar caused context menus to automatically
close themselves when the bar switches to a hidden state.

Unfortunately, SwiftUI doesn't have any methods to check if a context
menu is showing or not. A UIKit context menu had to be added. MacOS
doesn't have this problem, so add a view which uses a native or the
custom context menu depending on the user's device.

Signed-off-by: kingbri <bdashore3@proton.me>